requirements.txt
Flask==2.0.1
openai==0.26.4
python-dotenv==0.19.2
azure-cognitiveservices-speech==1.24.2